+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * Middle mouse drag now pans the timeline
     * There is an icon on the timeline to quickly toggle keyframes
     * Buttons to jump to the next/previous keyframe in the timeline
+    * Improved LottieFiles import dialog
 * Misc:
     * Switched to an even/odd version numbering scheme
 * Bug Fixes: